Revert "PD #108959: show av no program message when stream was stop."
This reverts commit b95320a88a4e8e00955c1965586f9e5ce08d13d9.

+++ am_adp/am_aout/am_aout.c
@@ -45,7 +45,7 @@ static AM_AOUT_Device_t aout_devices[] =
 /**\brief 根据设备号取得设备结构*/
 static AM_INLINE AM_ErrorCode_t aout_get_dev(int dev_no, AM_AOUT_Device_t **dev)
 {
-	if((dev_no<0) || (((size_t)dev_no)>=AOUT_DEV_COUNT))
+	if((dev_no<0) || (dev_no>=AOUT_DEV_COUNT))
 	{
 		AM_DEBUG(1, "invalid aout device number %d, must in(%d~%d)", dev_no, 0, AOUT_DEV_COUNT-1);
 		return AM_AOUT_ERR_INVALID_DEV_NO;
